Move shared parser helpers
