+++ src/datamodel_code_generator/base_config.py
@@ -21,6 +21,7 @@ from datamodel_code_generator._format_types import (
     PythonVersion,
     PythonVersionMin,
 )
+from datamodel_code_generator._shared_types import LiteralType  # noqa: TC001 - used at runtime by Pydantic
 from datamodel_code_generator.enums import (
     DEFAULT_SHARED_MODULE_NAME,
     AliasGenerator,
@@ -46,7 +47,6 @@ from datamodel_code_generator.enums import (
     UnionMode,
     VersionMode,
 )
-from datamodel_code_generator.parser import LiteralType  # noqa: TC001 - used at runtime by Pydantic
 
 if TYPE_CHECKING:
     from datamodel_code_generator.preset_names import PresetName as PresetNameValue
